+++ RELEASES.md
@@ -6,6 +6,8 @@
   `assign_str*` methods.
 * `Integer::gcd_coeffs_ref` now supports computing only one
   coefficient.
+* `Integer::invert_ref` and `Integer::pow_mod_ref` now return an
+  `Option`, not an object that is assignable to `Result`.
 
 Version 0.9.3 (2018-01-09)
 ==========================